refactor: cleanup topnav and restore layout suspense boundaries
