+++ components/TopNav/TopNavStaticContent.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import ProductPopoverContent from './ProductPopoverContent'
+import ResourcesPopoverContent from './ResourcesPopoverContent'
+import AuthButtons from './AuthButtons'
+
+export default function TopNavStaticContent() {
+  return (
+    <div>
+      <ProductPopoverContent />
+      <ResourcesPopoverContent />
+      <AuthButtons />
+    </div>
+  )
+}
